+++ smoove.go
@@ -1,3 +1,3 @@
 package smoove
 
-const Version = "0.1.5"
+const Version = "0.1.6-dev"